@@ -20,9 +20,9 @@ import static net.utp4j.data.UtpPacketUtils.RESET;
 import static net.utp4j.data.UtpPacketUtils.STATE;
 import static net.utp4j.data.UtpPacketUtils.SYN;
 import static net.utp4j.data.UtpPacketUtils.joinByteArray;
-import static org.junit.Assert.assertEquals;
+import org.junit.jupiter.api.Test;
 
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.*;
 
 public class UtpPacketUtilsTest {
 	
