@@ -14,12 +14,11 @@
 */
 package net.utp4j.channels.impl.alg;
 
-import static org.junit.Assert.assertEquals;
-
-import org.junit.Test;
-
 import net.utp4j.channels.impl.alg.MinimumDelay;
 import net.utp4j.channels.impl.alg.UtpAlgConfiguration;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class MinDelayTest {
 
